fix: address code review findings - cdn_ports u16, single-loop mixed strategy, acme cleanup
- types: cdn_ports() returns &[u16] per plan spec
- xhttp: single-loop mixed strategy (cdn port first, random fallback per
  iteration) instead of two-phase; remove redundant as-u16 casts
- acme: configured_provider_from_configs uses direct then_some instead of
  single-element iterator
- dispatch: save/restore lang in stale_domain_provider_callback test to
  prevent i18n global state leak

+++ rust/aegis/src/core/types.rs
@@ -25,7 +25,7 @@ impl DnsProvider {
         }
     }
 
-    pub const fn cdn_ports(self) -> &'static [i32] {
+    pub const fn cdn_ports(self) -> &'static [u16] {
         match self {
             Self::Cloudflare => &[443, 8443, 2053, 2083, 2087, 2096],
             Self::Route53 => &[],
